+++ domain/src/test/java/org/oppia/android/domain/classify/AnswerClassificationControllerTest.kt
@@ -24,13 +24,13 @@ import org.oppia.android.app.model.RuleSpec
 import org.oppia.android.app.model.SubtitledHtml
 import org.oppia.android.app.model.Translation
 import org.oppia.android.app.model.WrittenTranslationContext
-import org.oppia.android.domain.classify.InteractionObjectTestBuilder.createFraction
-import org.oppia.android.domain.classify.InteractionObjectTestBuilder.createMixedNumber
-import org.oppia.android.domain.classify.InteractionObjectTestBuilder.createNonNegativeInt
-import org.oppia.android.domain.classify.InteractionObjectTestBuilder.createReal
-import org.oppia.android.domain.classify.InteractionObjectTestBuilder.createSetOfTranslatableHtmlContentIds
-import org.oppia.android.domain.classify.InteractionObjectTestBuilder.createString
-import org.oppia.android.domain.classify.InteractionObjectTestBuilder.createTranslatableSetOfNormalizedString
+import org.oppia.android.app.player.state.testing.InteractionObjectTestBuilder.createFraction
+import org.oppia.android.app.player.state.testing.InteractionObjectTestBuilder.createMixedNumber
+import org.oppia.android.app.player.state.testing.InteractionObjectTestBuilder.createNonNegativeInt
+import org.oppia.android.app.player.state.testing.InteractionObjectTestBuilder.createReal
+import org.oppia.android.app.player.state.testing.InteractionObjectTestBuilder.createSetOfTranslatableHtmlContentIds
+import org.oppia.android.app.player.state.testing.InteractionObjectTestBuilder.createString
+import org.oppia.android.app.player.state.testing.InteractionObjectTestBuilder.createTranslatableSetOfNormalizedString
 import org.oppia.android.domain.classify.rules.algebraicexpressioninput.AlgebraicExpressionInputModule
 import org.oppia.android.domain.classify.rules.continueinteraction.ContinueModule
 import org.oppia.android.domain.classify.rules.dragAndDropSortInput.DragDropSortInputModule
